Check if ProjectProto is installed on microSD card
One of the key features of ProjectProto is the creation of a vendor partition
on the microSD card. Since Samsung Galaxy MSM8916 devices do not support a
vendor partition, check if the maximum partition count is 29 (on J5 2016 the
stock maximum partition count is 28) and if a vendor partition exists to
confirm that ProjectProto is installed on the microSD card already.