Validate eMMC partition layout
Ensure that the target partitions (system, cache, hidden, userdata) are in the
correct order and that no further partitions are followed after userdata
partition before the cloning process. Also store the IDs of the target
partitions and the system start sector. This is needed to post check if the
clone process was successfully.