Check system and vendor sizes to verify ProjectProto installation
Previously projectProtoInstalled checked the max partition count and the
presence of the vendor partition in the sgdisk --print output. For a more
robust and MSM8916-device-independent approach it now verifies the sizes of the
system and vendor partitions instead. This handles the edge case of Samsung
MSM8916 devices that have fewer partitions due to the absence of an FRP
partition.